requirements.txt
aniso8601==7.0.0
asgiref==3.5.0; python_version >= "3.7" and python_version < "4.0"
attrs==21.4.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
dj-database-url==0.5.0
django-cachalot==2.5.0
django-configurations==2.3.2; python_version >= "3.6" and python_version < "4.0"
django-debug-toolbar==3.2.4; python_version >= "3.6"
django-extensions==3.1.5; python_version >= "3.6"
django==3.2.12; python_version >= "3.6"
djangorestframework-api-key==2.1.0; python_version >= "3.6"
djangorestframework-simplejwt==5.0.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
drf-spectacular==0.21.2; python_version >= "3.6"
graphene-django-optimizer==0.8.0
graphene-django==2.15.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
inflection==0.5.1; python_version >= "3.6"
jsonschema==4.4.0; python_version >= "3.7"
promise==2.3
psycopg2-binary==2.9.3; python_version >= "3.6"
pyjwt==2.3.0; python_version >= "3.7"
pyrsistent==0.18.1; python_version >= "3.7"
pytz==2021.3; python_version >= "3.7" and python_version < "4.0"
pyyaml==6.0; python_version >= "3.6"
rx==1.6.1
singledispatch==3.7.0; python_version >= "2.6"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "2.6"
sqlparse==0.4.2; python_version >= "3.7" and python_version < "4.0"
text-unidecode==1.3
uritemplate==4.1.1; python_version >= "3.6"
